split up the Apl file into a few pieces
